// src/app/core/api/openweather-api.model.ts
// Generated by https://quicktype.io

export interface OpenWeatherApiResponse {
  lat: number;
  lon: number;
  timezone: string;
  timezone_offset: number;
  current: Current;
  hourly: Current[];
  daily: Daily[];
}

export interface Current {
  dt: number;
  sunrise?: number;
  sunset?: number;
  temp: number;
  feels_like: number;
  pressure: number;
  humidity: number;
  dew_point: number;
  uvi: number;
  clouds: number;
  visibility: number;
  wind_speed: number;
  wind_deg: number;
  weather: Weather[];
  snow?: Snow;
  pop?: number;
  rain?: Rain;
}

export interface Snow {
  '1h': number;
}
export interface Rain {
  '1h': number;
}

export interface Weather {
  id: number;
  main: string;
  description: string;
  icon: string;
}

export interface Daily {
  dt: number;
  sunrise: number;
  sunset: number;
  temp: Temp;
  feels_like: FeelsLike;
  pressure: number;
  humidity: number;
  dew_point: number;
  wind_speed: number;
  wind_deg: number;
  weather: Weather[];
  clouds: number;
  pop: number;
  rain?: number;
  snow?: number;
  uvi: number;
}

export interface FeelsLike {
  day: number;
  night: number;
  eve: number;
  morn: number;
}

export interface Temp {
  day: number;
  min: number;
  max: number;
  night: number;
  eve: number;
  morn: number;
}

// Generated by https://quicktype.io

export interface AirPollutionApiResponse {
  coord: Coord;
  list: List[];
}

export interface Coord {
  lon: number;
  lat: number;
}

export interface List {
  main: Main;
  components: { [key: string]: number };
  dt: number;
}

export interface Main {
  aqi: number;
}